FF net's hidden layers can now be manually chosen. Reordered input parameters
