+++ plugins/qnn/tasks.py
@@ -200,6 +200,9 @@ def calculation_task(self, db_id: int) -> str:
     postprocess_layers = [
         int(el) for el in input_params.postprocess_layers.split(",") if el != ""
     ]
+    hidden_layers = [
+        int(el) for el in input_params.hidden_layers.split(",") if el != ""
+    ]
 
     model_parameters = dict(
         n_qubits=n_qubits,
@@ -209,6 +212,7 @@ def calculation_task(self, db_id: int) -> str:
         output_size=n_classes,
         preprocess_layers=preprocess_layers,
         postprocess_layers=postprocess_layers,
+        hidden_layers=hidden_layers,
         single_q_params=(weights_to_wiggle != 0),
     )
